Fix shape and length reporting of gridded interfaces for Python 3.6 on Win 64 (#2903)

+++ holoviews/core/data/image.py
@@ -79,7 +79,7 @@ class ImageInterface(GridInterface):
 
     @classmethod
     def length(cls, dataset):
-        return np.product(dataset.data.shape[:2])
+        return np.product(dataset.data.shape[:2], dtype=np.intp)
 
 
     @classmethod
